Move FrameBound tests into the package of the FrameBound class
