+++ sql-parser/src/test/java/io/crate/sql/tree/OffsetFollowingFrameBoundTest.java
@@ -20,21 +20,21 @@
  * agreement.
  */
 
-package io.crate.execution.engine.window;
+package io.crate.sql.tree;
 
-import io.crate.test.integration.CrateUnitTest;
 import org.junit.Before;
 import org.junit.Test;
 
 import java.util.Comparator;
 import java.util.List;
 
-import static io.crate.sql.tree.FrameBound.Type.PRECEDING;
+import static io.crate.sql.tree.FrameBound.Type.FOLLOWING;
 import static io.crate.sql.tree.WindowFrame.Mode.RANGE;
 import static io.crate.sql.tree.WindowFrame.Mode.ROWS;
 import static org.hamcrest.core.Is.is;
+import static org.junit.Assert.assertThat;
 
-public class OffsetPrecedingFrameBoundTest extends CrateUnitTest {
+public class OffsetFollowingFrameBoundTest {
 
     private Comparator<Integer> intComparator;
     private List<Integer> partition;
@@ -46,14 +46,14 @@ public class OffsetPrecedingFrameBoundTest extends CrateUnitTest {
     }
 
     @Test
-    public void test_preceding_start_in_range_mode() {
-        int frameStart = PRECEDING.getStart(RANGE, 0, 4, 4, 2L, 4, intComparator, partition);
+    public void test_following_end_in_range_mode() {
+        int frameStart = FOLLOWING.getEnd(RANGE, 1, 5, 1, 2, 4, intComparator, partition);
         assertThat(frameStart, is(3));
     }
 
     @Test
-    public void test_preceding_start_in_rows_mode() {
-        int frameStart = PRECEDING.getStart(ROWS, 0, 4, 3, 2L, null, intComparator, partition);
-        assertThat(frameStart, is(1));
+    public void test_following_end_in_rows_mode() {
+        int frameStart = FOLLOWING.getEnd(ROWS, 1, 5, 1, 2L, null, intComparator, partition);
+        assertThat(frameStart, is(4));
     }
 }
